Function parses patient gender
Form can now parse the gender  and species checked by the user and adds correct pronouns and species-specific text as necessary.
Began working on creating function that can see when one of the patient's object keys are null and then fill it in with the appropriate default statement  in the output.