Webpage takes old history and outputs new history
Finished (for now)! Works as expected thus far. Going to add a bit more styling later, but don't have the imaging editing programs I want to use on my current computer.
Still hoping to add an "append" button that can add missing values to outputs, but I'm publishing the page as is for now.